+++ app/x2mail/v/current/app/libraries/X2Mail/Engine/Actions/Pgp.php
@@ -368,23 +368,24 @@ trait Pgp
 					*/
 
 					$summary = \defined('GNUPG_SIGSUM_VALID') ? [
-						GNUPG_SIGSUM_VALID => 'The signature is fully valid.',
-						GNUPG_SIGSUM_GREEN => 'The signature is good but one might want to display some extra information. Check the other bits.',
-						GNUPG_SIGSUM_RED => 'The signature is bad. It might be useful to check other bits and display more information, i.e. a revoked certificate might not render a signature invalid when the message was received prior to the cause for the revocation.',
-						GNUPG_SIGSUM_KEY_REVOKED => 'The key or at least one certificate has been revoked.',
-						GNUPG_SIGSUM_KEY_EXPIRED => 'The key or one of the certificates has expired. It is probably a good idea to display the date of the expiration.',
+						GNUPG_SIGSUM_VALID => 'The signature is valid.',
+						GNUPG_SIGSUM_GREEN => 'The signature matches the message.',
+						GNUPG_SIGSUM_RED => 'The signature is invalid: it does not match the message.',
+						GNUPG_SIGSUM_KEY_REVOKED => 'The signing key has been revoked.',
+						GNUPG_SIGSUM_KEY_EXPIRED => 'The signing key has expired.',
 						GNUPG_SIGSUM_SIG_EXPIRED => 'The signature has expired.',
-						GNUPG_SIGSUM_KEY_MISSING => 'Can’t verify due to a missing key or certificate.',
-						GNUPG_SIGSUM_CRL_MISSING => 'The CRL (or an equivalent mechanism) is not available.',
-						GNUPG_SIGSUM_CRL_TOO_OLD => 'Available CRL is too old.',
+						GNUPG_SIGSUM_KEY_MISSING => 'The public key of the signer is missing.',
+						GNUPG_SIGSUM_CRL_MISSING => 'The revocation list is not available.',
+						GNUPG_SIGSUM_CRL_TOO_OLD => 'The revocation list is too old.',
 						GNUPG_SIGSUM_BAD_POLICY => 'A policy requirement was not met.',
-						GNUPG_SIGSUM_SYS_ERROR => 'A system error occurred.',
+						GNUPG_SIGSUM_SYS_ERROR => 'The signature could not be checked because of a system error.',
 //						GNUPG_SIGSUM_TOFU_CONFLICT = 'A TOFU conflict was detected.',
 					] : [];
 
 					// Verified, so no need to return $result['text'] and $result['signature']
 					$result = [
 						'fingerprint' => $info['fingerprint'],
+						'keyid' => $info['keyid'] ?? '',
 						'validity' => $info['validity'],
 						'status' => $info['status'],
 						'summary' => $info['summary'],